Move native intrinsics to a separate class for clarity
